Play with time intervals
This patch adds necessary code to define time intervals and check
whether two intervals overlap or not. In addition, it uses QuickCheck
to verify that the overlap function is both reflexive and
symmetric. Additional tests have to be written, but these two
properties have helped to find several missing cases in the
implementation.