Introduce the notion of event in ogma model
This patch defines a very important data type: Event. In addition to
that, it uses the previously defined functions overlap and collide to
define a powerful filtering function. With that, it is only a matter
of good UX to be able to browser in a convenient way what happened in
a story.